requirements.txt
delta-spark
pyspark
fastapi
uvicorn
xgboost
tensorflow
pandas
numpy
scikit-learn
pyarrow
fastparquet
requests
